requirements.txt
mysql-connector-python==8.0.21
python-docx==0.8.10
